Load AngularJS, FontAwesome, JQuery and Bootstrap files (js and css assets) locally, instead of from a CDN.
This should help the problem with users from China not being able to load JS files from the google CDN.

These assets are minified and compiled into a single file by the asset pipeline. This means the first load will be unavoidably large, but from then on it should be cached in the browser. Cloudflare should work as an almost-CDN serving files compressed and from a geographically close server, supposedly.