When an entry is closed, all tooltips in the entry (entry title link, share button) are hidden.
This fixes the following issue: when the user clicks on the title link, a new tab opens. This does not trigger the mouseover event and the tooltip stays visible until the user clicks somewhere else. If the user uses the mouse to navigate this is not a problem because he has to click to open another entry, so tooltips will be hidden; but since implementing keyboard shortcuts it is possible that the user will open another entry, closing the one with the tooltips, using a keyboard shortcut; this does not trigger any mouse event, and the result was that the tooltip for the title of the closed entry was still visible, which looks really bad.

The fix is to forcibly hide tooltips in an entry when it is closed.