Use asset path helpers inside FontAwesome's SCSS files to determine the actual path of each font file.
This eliminates the need to have a separate SCSS file specifying these paths. It's better to define the paths for the FontAwesome font just once, in the proper place under the bower-managed font-awesome folder.

In the future there may be font-awesome updates that may require to do this again manually.