When updating failing_since and available attributes of a feed as part of a scheduled update, first reload the feed from the db.
This should help with rare errors in which the feed object had been frozen and trying to change it raised an error. Also a control has been added to make sure the feed still actually exists in the db, in case it has been deleted by the time the execution reaches that point.