Rescue OpenSSL::SSL::SSLError when updating a feed, as one of the possible errors that can happen which do not mean the job itself has failed.
If this error (and several others) happens, this means that the connection to the feed has failed. As far as sidekiq is concerned the job is successful, so the error is rescued to prevent sidekiq from putting the job in the "retries" queue.

This error means that the SSL negotiation with the feed server has failed. If this happens during a scheduled update, the feed is marked as failing and eventually as permanently unavailable, as usual.