The uniqueness validation for the URLs of feeds and opml import failures is case sensitive.
URLs in general are case sensitive. In most cases the server is actually case insensitive, but the spec mandates dealing with them in a case sensitive way.

This should fix a bug in which trying to subscribe to a feed that already exists in the db, but with different case in some characters, fails with an AR uniqueness validation error.