Wherever a unit test asserts that an error should be raised, specify the error class expected.
Recent rspec versions issue a warning if the following is done without passing an error class:

expect {...}.to raise_error

The reason is that this way any error is captured, which can mask programming errors with missing arguments and such.

The fix is to pass the expected error class as argument to raise_error.