When moving a feed to a new folder during acceptance testing, introduce a small delay between the popup appearing and entering the new folder name.
This is necessary to give the client time enough to disable the keyboard shortcuts service; this service must be disabled or letters associated with a shortcut cannot be used as part of the folder name.

Older selenium versions worked without the delay, it seems a recent update has made selenium so fast that unless this delay is introduced, the client does not have time to disable the shortcuts service before the test tries to enter the new folder name.