Added Net::HTTPBadResponse to the list of expected errors that can be raised during a scheduled feed update.
If this error is raised, it is rescued so that the job itself does not fail, preventing Resque from retrying it. The feed is marked as failing and eventually as permanently unavailable, as usual.

This error sometimes happens in production because of a wrong chunk size reported by the server:

	Net::HTTPBadResponse: wrong chunk size

Also the list of expected and rescued errors in all workers that deal with HTTP connections has ben homogeinized as much as possible (some error classes were missing from some rescue clauses).