Demo user cannot be deleted if the demo user flag is enabled.
This commit adds a control at the model level so that the app does not enqueue a delete job for the demo user nor does it try to lock it as soon as the user requests a profile deletion.

There are already controls at the AR callback level that will stop the deletion; however the DestroyUserWorker instance will be enqueued, which is a waste of resources, and it will fail with an error because of the failure in the before_destroy AR callback. Worse, Sidekiq will put it in the retries queue, which is even more wasteful.

It's better to stop all this from happening when User#delete_profile is invoked. No worker is enqueued and the locking attempt is not performed, if the user in question is the demo user.