Make the uniqueness constraint on the guid attribute of Entry and DeletedEntry case-sensitive.
It makes sense that the guid is case-sensitive; often the guid has the same value as the entry URL, and URLs are case sensitive.

This means that e.g. an entry with guid "AAA" and another with guid "aaa" will be regarded as different entries and both will be saved in the db; the old behavior (case-insensitive uniqueness constraint) meant that the app considered both to be the same entry and only one would be saved in the db.

An important side effect of this is that when searching in the deleted_entries table, the SQL condition will not have "LOWER(guid)=..." but instead "guid=...". This means that the SELECT will be able to use the (guid, feed_id) index on the deleted_entries table. This should be a huge performance boost.