Fixed Sidekiq version to 3.4.2
The latest Sidekiq 3.5.0 version depends on Celluloid 0.17.0. However Sidetiq (which we use for recurring jobs) is incompatible with this Celluloid version, for more details see:

  https://github.com/tobiassvn/sidetiq/issues/140
  https://github.com/tobiassvn/sidetiq/issues/141

This makes Sidekiq (using Sidetiq) crash as soon as it's started.

The fix is to use for now the older Sidekiq 3.4.2, which depends on Celluloid 0.16.0 which in turn is compatible for Sidetiq.

Sidetiq doesn't seem to be actively maintained, an official patch may not happen. Most people suggest migrating to sidekiq-cron.