Removed "visible: true" during tests when clicking on a feed.
It seems Selenium-webdriver, since 2.48.1, does not work correctly if the "visible: true" argument is passed to "find" in acceptance tests that click on a sidebar feed link. The link is visible, the "find(..).click()" call does not raise errors, but the link is not actually clicked, which makes the rest of the test fail.

For now I'm removing the argument from the "find(..)" call, so that the element to be clicked is found regardless of being visible or not. I'm not too happy with this but the way the tests are written links should always be visible when the execution reaches that point, anyway.