Added Zlib::DataError to the errors that can be expected to be raised when fetching a feed.
If this (rare) error happens, this means that something is wrong with the zipping of the HTTP response. The error is not raised, to avoid sidekiq adding the worker to its list of retries; rather, the feed is marked as "currently failing" and eventually is marked as unavailable, same as other possible errors that indicate that something's wrong with the feed itself, not with the worker fetching it.