Fix displaying the number of unread entries in the favicon for Chrome.
Currently favico.js has a bug which only updates the last element with rel="icon" if there are several in the page head. This maeans that the favicon must be last in the head, particularly it must appear after the 192x192 icon (used in the mobile manifest).

Related to this favico.js bug, might become unnecessary when it is fixed:

  https://github.com/ejci/favico.js/issues/85