Downgrade velocity.js 1.2.3 -> 1.2.2
Velocity.js 1.2.3 has broken passing an offset and a container to a scroll animation. When autoscrolling the sidebar after opening a folder, version 1.2.2 reliably scrolls the folder to the top of the sidebar, while version 1.2.3 scrolls the folder to some point that seems to depend on viewport height, and can even place the folder outside the viewport.

Docs and commits do not say anything about this change. It seems a regression, but it's hard to pin down the guilty commit.