WIP: when a feed scheduled update fails, attempt autodiscovery only if it's been failing for a (configurable) time, instead of immediately.
This helps with the case in which autodiscovery is too zealous, and performs autodiscovery on a feed that fails for a very short time. If the website supports multiple feeds, this could easily lead to the feed fetch_url being replaced with that of a different feed from the page; in this case retrying the update a few times instead of triggering autodiscovery would have probably recovered the feed.

The above situation (too-zealous autodiscovery) is an error that is hard for the user to detect and even harder to correct (non-admins have no other option but to unsubscribe and resubscribe to the correct feed). Automatic autodiscovery is useful in some cases, but it's better to wait until after the feed has been failing for some time, to make sure it's not an intermittent but recoverable failure.