Included fastclick 1.0.6 library, managed by Bower, along with its initialization.
Angular-touch 1.5 will deprecate and disable by default the removal of the 300ms delay in touchscreens. They recommend using fastclick or Tappy instead. This commit prepares for when angular 1.5 is included.

The reason for the deprecation is that the ng-touch version of ngClick has never worked properly, causing issues with phantom touch events. Also browsers are smarter these days, imposing the 300ms delay in few cases (but fastclick.js is still necessary, e.g. for homescreen web apps in iOS). For more about this deprecation see:

  https://github.com/angular/angular.js/issues/6251
  https://github.com/angular/angular.js/commit/0dfc1dfebf26af7f951f301c4e3848ac46f05d7f
  http://developer.telerik.com/featured/300-ms-click-delay-ios-8/