Changing the state (read/unread) of a single entry is performed in a web worker.
This means the main thread is not used for this ajax call, which makes the UI a bit more responsive, specially in multicore devices.

This change does not improve performace a lot, because this ajax call is not resource intensive. However it will be the blueprint for moving other processes to web workers (at the very least, all ajax calls, both scheduled and user initiated).

Also the new architecture based on web workers has allowed implementing a new feature: if there is a network failure during the ajax call, instead of immediately showing an error to the user the ajax call is reattempted a number of times, before giving up. This hides intermittent network failures from the user.