When clicking on a feed link in an entry header, immediately switch to the feed.
The old behavior was to mark the open entry as unread and only then switch to the feed. This meant that the open entry appeared in the entries list, but there could be a delay until the response indicating a successful state change was received.

With this change the switch to the feed is immediate but the entry that was open stays in the same state; this means that if it was read, it won't appear in the entries list by default. It may actually be a good thing, the old behavior was too "magical".