Changed keypress management to an angularjs ng-keypress directive.
This means that keyboard shortcut handling is performed in an angularjs service triggered by ng-keypress. This way the UI is immediately updated (e.g. entries are immediately marked as read, etc) without the need for an ugly $roootScope.apply().

Unfortunately keyboard shortcuts are again bound to lowercase characters. Angularjs does not seem to have a method in the keypress event to recover the case-insensitive keycode.