If a 404 is returned when querying for the state of a single refresh or subscribe job, do not show an error message.
This prevents an error being displayed in the following scenario: user has a running job (the client periodically queries for its state), then there is a network connectivity failure, and while the network is down the user dismisses the job status alert (which sends an HTTP DELETE). When network connectivity returns, the DELETE goes through as well as a GET querying for the job state; the GET returns a 404, because we've just deleted that job status record. In this case we don't want the user to see an error message, there is no error.