Fixed showing an alert message when a running export job finishes.
Introduced a boolean flag in the service, instead of trusting the flag that indicates whether to show the alert to be passed to the service back from the web worker. There should never be two parallel invocations to the service, so the boolean flag should work without issue, even though I admit it's not the most elegant code.