Explicitly return service object from userDataSvc.
After the addition of the web worker code, I'm not sure if the worker variable is visible from outside the service when using the implicit service return. It's better to encapsulate the service functions in an object and explicitly return it, so that the worker variable is private to the service.