WIP: pass an "operation" argument to web workers.
This argument is not passed to the server via Ajax; rather, it is returned untouched to the main thread via callback when the Ajax call returns.

It serves for the main thread to know which operation has just finished when the callback is called, and act accordingly. This way angular services can use a single worker for multiple purposes (operations) without having to launch several web workers with the performance overhead that incurs.