Fixed bug when successfully adding a subscription.
Adding a subscription no longer takes the user to the Start page. This could be surprising for the user. Also no alert is displayed when the subscribe job has been enqueued, only once it finishes.
Instead job states are loaded in the background: this means that once the subscribe job finishes, either successfully or with a failure, a temporary alert will be displayed at the bottom of the screen, regardless of where in the app is the user. This feels more natural.

Also a small delay has been added in acceptance tests when opening the feeds menu, to give enough time for the menu to fully open before clicking on any of its entries.