When adding a subscription, the user is shown the start page.
This was removed but now is added back. Showing the start page is desirable because the user will get feedback on his action, in the form of a progress alert. Removing this would mean that the user would get no feedback until the import worker finishes, which can take some time; this could lead the user to think that the app has not received the subscribe command.