Ignore entries received if they are not the expected ones.
Loading entries can take some time. During this time the user can click on several sidebar entries or folders. The expected result is to display the entries for the last clicked sidebar link, any other entries received (from the previous Ajax calls) are ignored.

I'd prefer cancelling the ajax request itself, but xhr.abort() does not seem to work in chromium. I'll have to live with the overhead of completing ajax requests for entries the client is no longer interested in, and then ignoring them.

Fortunately the client is aware of the last clicked sidebar link, either feed or folder, and can ignore entries that do not correspond to the last requested feed/folder and page.