More correct behavior for the loading_entries_busy flag.
We no longer check this flag before loading entries. It is entirely possible that e.g. while entries for a feed are loading, the user clicks on a different feed. In this case we do want to load entries for the second feed, and ignore the response with the entries for the first feed.

Also resetting the entries pagination no longer sets the loading_entries_busy flag to false. Same as above, it is possible that while entries for a feed are loading (loading_entries_busy is true) the user clicks on a different feed. This triggers a reset of the entries pagination (page is set to zero, entries list is emptied etc) to prepare for loading entries for the second feed; but this does not mean the app is not busy, in fact entries loading continues to be busy until the response with the entries for the second feed is received. This means that resetting entries pagination should not necessarily set this flag to false.

As a side effect, when clicking on the start link in the sidebar this flag must be explicitly set to false from startPageSvc, because it is no longer done in entriesPaginationSvc.