Navbar button animation now works every time the button is clicked.
The problem was that after the animation ran once, the background opacity was set to 0. This meant that trying to animate the button again had no visible effect, because the animation changed the background color which was no longer visible (totally transparent).

The fix is to set the background color to the same as the navbar, and its opacity to 1, as soon as the animation ends. This way the same animation can run again with the same visible effect.