WIP: load feeds in a web worker.
Moved code that runs when the web worker finishes returning an error to separate functions for each of the possible operations the worker can perform (load a single feed, load a page of all feeds, or load feeds in a folder).

The web worker callback code is much more readable now.