When opening folders in acceptance tests, wait a bit for the animation to open.
This is necessary in recent selenium versions, which are more strict checking if an element is clickable. If we order the check to click on a link under a folder without waiting for the folder open animation to finish, it's likely that the feed won't yet be visible, so the click will fail and the test will fail as well.