Window size during acceptance testing is much bigger.
This avoids having to scroll a lot to make elements visible and clickable (recent selenium versions are a bit anal about not being able to click on elements outside the viewport).

Also fix an error where the wrong value was being entered info a form field. The test still passed because the pass condition was being unable to log in, but the way it was written the login would always fail.