The worker that resets the demo user also updates unread count of his subscriptions.
Previously, all entries were marked as unread but the subscriptions' unread count was not updated. This meant that even though the entries were unread, anyone who logged in as the demo user could see a wrong unread count in the sidebar; it was even possible to see no feeds in the sidebar (all feed counts set to zero) even though all entries were actually marked as unread. This error was temporary because the next scheduled update of each feed would update the unread count with the correct number.

However it's better if when the ResetDemoUserWorker runs, after marking entries as unread, it also updates the unread count of all subscriptions. This way the reset of the demo user is complete, not half-done.