Added explicit dependency in Gemfile from inherited_resources gem.
It is necessary to use a yet untagged version from github so activeadmin is compatible with rails 5.

I would like to remove both inherited_resources (dev considers it deprecated and out of maintenance) and activeadmin (stable version releases happen at a glacial pace). Maybe switch to another admin interface in the future.