Updated rails 4.2.7 -> 5.0.0. Added dependencies on github master of sinatra, rack-protection gems.
There is currently no released version of sinatra, rack-protection that is compatible with rails 5 (specifically, the declared dependency version on rack is a problem). For now I'm using their github master, when a stable version is released I'll switch to it. The explicit dependency on these two gems will probably be removed (sidekiq has a dependency on sinatra, no need for it to be in the Gemfile directly).