WIP: handle special feeds (e.g. demonoid) that need special handling.
This is intended for feeds which are actually faulty or somehow break spec, but we want feedbunch users to be able to subscribe to them anyway; in some cases, whatever is wrong with the feed can be corrected when feedbunch fetches it. E.g. demonoid feeds have entry guids that change over time, resulting in duplicated entries. The pattern with which those guids are generated however is predictable and can be fixed in our side.

This is of course brittle and could break if the feed owners change their feed generation algorithm in the future. This feature should only be used in special cases, after trying to contact feed owners and convince them to fix their feeds.