Special handling of Demonoid feeds.
The entry guids that change over time are replaced with guids that are permanent. This works for now but is brittle, it could break if Demonoid changes the algorithm they use to generate guids.

The foundation is there to add special handling for other feeds. The necessary steps would be:

- add a new entry in the special_feeds.yml file indicating the hostname of the feed and the class that will implement the handling.
- implement a new class that implements the handle_entry(entry) class method, making the necessary changes in the before_validation callback of the entry.