Use rails helpers for the font paths in fontawesome SCSS, instead of hardcoded paths.
The paths used by default by fontawesome do not exist in a rails app, the asset pipeline creates different paths. Using rails helpers is a must to find out the actual asset path for each font.

There may be fontawesome updates in the future for which this change has to be manually applied again.