Fixed sending emails from Devise (confirmation, unlock, invitation and reset password)
After migrating to rails 5 autoloading files in lib has been disabled by default. This means that the MailerButtonStyler class, which is used to get inline CSS for the action button in mails sent by Devise, is not defined in the mailer view in the production environment.

The easiest fix is to require the class in the four mailer views that use it. A bit ugly but less unintended consequences than messing with autoloading in the app.