Fix issues with feeds that have changed URL when resetting the demo user.
Before this commit if one of the feeds in the default list for the demo user changed URL (with a redirect to the new URL), the worker crashed. This is because it tried to reset unread entries of the feed using the url in the configuration to find it, and that feed no longer exists (its url has been changed to a new one automatically the first time it was fetched after the URL change).

The fix is to get a list of the feeds actually subscribed by the demo user after resetting subscriptions, and then using this list both for resetting unread entries and for resetting feed/folder associations. The list in the config may be obsolete, but if there is a redirection in place to the new URL, this will be transparent for the end user.