Pinning redmon gem dependency to version 0.0.10
Version 0.0.11 does not declare its dependencies correctly. Because of this necessary gems are missing from the bundle, which causes app startup to fail (specifically because the haml gem is missing, but even if I include haml in my Gemfile other errors will arise later because of other missing gems).

See https://github.com/steelThread/redmon/issues/86

Awaiting for my PR to redmon fixing this to be merged and included in a tagged version, before unpinning redmon dependency in the Gemfile.