Removed 'style' from the list of allowed elements in the relaxed sanitizer configuration.
This relaxed config is used when sanitizing entry contents and summaries, which have to let some html pass. Other entry attributes use the strict sanitizer config, which already blocks style tags.

This change intends to stop style tags from being in entry content and summaries. This is because a style tag can contain CSS styles that clash with the app styles and make the app almost unusable as soon as the entry is loaded.

This even removes mismatched style tags. E.g. this is a real entry that before this change passed sanitization, but now the inline CSS is removed:

<p>some content</p>
<style>
div {
font: 18px serif;
margin-bottom: 2.5%;
background-color: #e1f5fe;
}</p>
<p>div.auto {
hyphens: auto;
}
</style>
<p>more content</p>