Manually apply an angularjs digest cycle when invitations_worker returns.
It seems web workers are outside the digest cycle and a digest must be manually invoked in the worker callback. Otherwise changes in the model (e.g. flags that make alerts temporarily visible) may not be reflected in the view (the alert may not become visible despite the flag becoming true).