Removed .erb extension from coffeescript file that didn't really need it.
We use .erb in the asset pipeline to enable the use of the asset_path helper in coffeescript files. This is very useful to get the actual URL of a web worker's javascript file, which we need to invoke the worker from the coffeescript code.

However ng-ShowHideReadSvc doesn't use web workers. It doesn't use any erb helpers either, which means the .erb extension in this case is not necessary and better removed (removing the erb treatment of this file from the asset pipeline will speed up asset compilation, even if only a minuscule amount).