When using keyboard shortcuts to select a sidebar link, hide any entry tooltips that are currently visible.
This fixes a bug in which entry tooltips could remain on the page after selecting a sidebar link with keyboard shortcuts (start-page, feed or read-all), even though the entry to which the tooltips belonged had disappeared from the page.

This is necessary because bootstrap tooltips react well to focus changes made with the mouse, but operating with keyboard tooltips does not fire mouse and focus change events, which means that bootstrap is not aware that the tooltip has to disappear; we have to manually force them to hide.