When checking if an entry already exists as part of a feed refresh, sanitize the entry guid before the check.
This is necessary because the Entry model sanitizes the guid (it is an outside input that cannot be trusted) and this can change the guid as it is in the fetched XML. E.g. if the guid in the XML has a "&amp;" sequence (e.g. inside a CDATA), what actually gets saved is "&"; this is an artifact of Loofah scrubbing that cannot be avoided.

So, when checking if the fetched entry already exists and therefore we can skip processing it, we must sanitize the guid before building the SQL query, otherwise in the above example we will not find the already existing entry and we will try to process an entry that already exists, getting an Active Record validation error just before saving.