Workaround to deal with itunes feeds that are incorrectly treated as feedburner feeds by feedjira.
These feeds end up with the enclosure-url in the entry.image attribute, weirdly enough. If the entry has no url and has an image attribute instead, default the url to the value of the image attribute.

This may open another category of bugs if the image attribute does not contain a valid URL...