Caused an error that would cause an entry to be skipped because it contained an img with src="http://"
The error was caused because trying to parse "http://" with the Addressable gem raises an error, because it is an invalid URL. The fix is to rescue the error and replace the invalid URL with an empty string, so that the image won't be loaded but the entry itself is correctly saved in the database. As an added bonus, this also fixes errors that could be caused by other invalid URLs in <img> tags in entries (I haven't seen any in production, but they could happen).