Fix bug in entries ordering.
The ordering of entries is determined by the server, clients assume the server is sending entries in the correct order. This is necessary so pagination does not do strange things.

Most of the time the published date is used to order entries. However in some cases several entries have the same publish date (this may happen in the feed xml or maybe publish dates are absent from the xml and Feedbunch has assigned them when the feed was fetched). In this case the created_at field is as a secondary criteria to order entries in a repeatable way.

The created_at date should be treated similar to the published date for this ordering: entries are ordered by the publish date in descending order, and any entries that have the same publish date are ordered by the created_at date in descending order.

The bug fixed is that, in some cases, entries that had the same publish date were ordered in ascending order by the created_at field. This had several side-effects: the ordering of entries was different depending on whether the "read all" button was checked or not; and if two entries with the same publish date were at the top of the entries list, clicking on the "read all" button didn't actually mark one of them as read, and that entry appeared again unread if the user reloaded the entries list.

Now the correct descending order by created_at is used in all cases. The impact was not big because this happens very rarely.