Fixed bug marking multiple entries as read.
When the user clicks on "read all", the server first decides which entries have to actually be marked as read. This is because new entries may have been fetched recently which the user has not yet loaded in the entries list, and in this case we do not want to mark these newer entries as read without user knowledge.

The server was ordering entries based on their publish date and only older entries were marked read. If several entries have the same publish date as the oldest entry currently loaded in the client, the id was being used as a secondary criteria, so only entries with lower ids were marked read.

This is an error. Actually the criteria here should be the same as when returning entries to the client: first order entries by publish date, then entries with the same publish date are ordered by the created_at field (also in descending order), and entries with the same both publish date and created_at date are ordered by id descending (this last case should not happen, but who knows).