Add user_id to an entry_states db index.
This commit complements 048175baeaae8c25b131c5f90a837e00bd150528. That
commit added user_id to the index that is used when retrieving unread
entries. This commit does the same but for the index that retrieves all
entries regardless of state. Both indexes are very similar, the
difference is that the first is a partial index (only records with
read=false are indexed) which is faster if retrieving unread entries. By
the way, partial indexes only work in postgresql I think.

This will speed up retrieving all entries regardless of state (when the
"show all entries" button is clicked).