Use the file backend for the rack-cache entity store.
The noop backend is causing problems. There is a feed
(http://oglaf.com/feeds/rss/) that for some reason is being treated as
"stale, invalid" (i.e. the http caching decides it must be
re-downloaded) but instead of returning the actual feed body to the
invoker, it receives a blank string instead. I cannot reproduce it in
development, in other enviroments different from production I'm getting
the expected "stale, valid" trace and therefore the body is irrelevant,
but in production there is something different that caused the trace to
have an "invalid" at the end despite returing an empty body.

I'll have to implement cron scripts to cleanup the hard drive and ensure
not too much space is taken up by the entity store, which never cleans
up itself. Maybe deleting entities older than a month.