Fixed acceptance tests.
Some tests were failing because they were attempting to click on the
sign up button while it was not visible. No idea why they passed before,
probably a recent change in selenium or chrome itself. Fix is to scroll
the page enough to show the signup button before attempting to click it.