Standalone mode in mobile browsers when launched from the homescreen.
A recent change in chrome 58 means that "fullscreen" for the display
attribute in the manifest hides the navbar and status bar in
smartphones. This is not intended, we don't actually need fullscreen for
feedbunch. The "standalone" value hides the browser navbar but still
shows the smartphone navigation buttons and status bar, which is much
better.